Warn when we select a LB IP that is not well defined.
When we get multiple IP addresses back for a LB, we have to pick one.
The choice of _which_ one is not well-defined, and might lead to odd
behavior if there are multiples!

We're logging it here, hopefully in the future if someone touches this,
they'll find the logs and adjust the method appropriately.